requirements.txt
dnspython3
click
